requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi @ file:///private/tmp/cffi-20231015-5876-1oubxev/cffi-1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
Django==5.0.6
djangorestframework==3.15.1
docutils @ file:///private/tmp/docutils-20231016-5777-1hh809k/docutils-0.20.1
Flask==3.0.2
frozenlist==1.4.1
hangups==0.4.18
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
MechanicalSoup==0.12.0
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
peewee==3.17.1
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.8
pycparser @ file:///private/tmp/pycparser-20231013-6010-ksuqao/pycparser-2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.8.2
pytz==2024.1
readlike==0.1.3
ReParser==1.4.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
twilio==9.0.0
tzdata==2024.1
urllib3==2.2.1
urwid==2.1.2
Werkzeug==3.0.1
yarl==1.9.4
